Refactor all js files into index.js and tests into index.tests.js
