--- reverse-string.js
@@ -1,10 +1,0 @@
-const reverseString = (str) => {
-  const splitStr = str.split('');
-  const reverseArr = splitStr.reverse();
-  const joinArr = reverseArr.join('');
-
-  return joinArr;
-};
-
-module.exports = reverseString;
-
